+++ cmd/config.go
@@ -5,57 +5,99 @@ import (
 	"io/ioutil"
 	"runtime"
 
-	"github.com/urfave/cli"
 	"gopkg.in/yaml.v3"
 )
 
 // Config represents what can be defined as a yaml config file
 type Config struct {
+	// GitLab configuration
 	Gitlab struct {
-		URL                string `yaml:"url"`                  // The URL of the GitLab server/api (default to https://gitlab.com)
-		Token              string `yaml:"token"`                // Token to use to authenticate against the API
-		HealthURL          string `yaml:"health_url"`           // The URL of the GitLab server/api health endpoint (default to /users/sign_in which is publicly available on gitlab.com)
-		DisableHealthCheck bool   `yaml:"disable_health_check"` // Whether to validate the service is reachable calling HealthURL
-		DisableTLSVerify   bool   `yaml:"disable_tls_verify"`   // Whether to skip TLS validation when querying HealthURL
-	}
+		// The URL of the GitLab server/api (default to https://gitlab.com)
+		URL string `yaml:"url"`
 
-	MaximumGitLabAPIRequestsPerSecond      int        `yaml:"maximum_gitlab_api_requests_per_second"`        // Maximum amount of requests per seconds to make against the GitLab API (default: 10)
-	ProjectsPollingIntervalSeconds         int        `yaml:"projects_polling_interval_seconds"`             // Interval in seconds at which to poll projects from wildcards
-	RefsPollingIntervalSeconds             int        `yaml:"refs_polling_interval_seconds"`                 // Interval in seconds to fetch refs from projects
-	PipelinesPollingIntervalSeconds        int        `yaml:"pipelines_polling_interval_seconds"`            // Interval in seconds to get new pipelines from refs (exponentially backing of to maximum value)
-	FetchPipelineJobMetrics                bool       `yaml:"fetch_pipeline_job_metrics"`                    // Whether to attempt to retrieve job metrics from polled pipelines
-	FetchPipelineVariables                 bool       `yaml:"fetch_pipeline_variables"`                      // Whether to attempt to retrieve variables included in the pipeline execution
-	PipelineVariablesFilterRegexp          string     `yaml:"pipeline_variables_filter_regex"`               // Regex to filter pipeline variables, deafaults to '.*'
-	OutputSparseStatusMetrics              bool       `yaml:"output_sparse_status_metrics"`                  // Whether to report all pipeline / job statuses, or only report the one from the last job.
-	OnInitFetchRefsFromPipelines           bool       `yaml:"on_init_fetch_refs_from_pipelines"`             // Whether to attempt retrieving refs from pipelines when the exporter starts
-	OnInitFetchRefsFromPipelinesDepthLimit int        `yaml:"on_init_fetch_refs_from_pipelines_depth_limit"` // Maximum number of pipelines to analyze per project to search for refs on init (default: 100)
-	DefaultRefsRegexp                      string     `yaml:"default_refs_regexp"`                           // Default regular expression to filter refs to fetch
-	MaximumProjectsPollingWorkers          int        `yaml:"maximum_projects_poller_workers"`               // Sets the parallelism for polling projects from the API
-	DisableOpenmetricsEncoding             bool       `yaml:"disable_openmetrics_encoding"`                  // Disable OpenMetrics content encoding in prometheus HTTP handler (default: false)
-	Projects                               []Project  `yaml:"projects"`                                      // List of projects to poll
-	Wildcards                              []Wildcard `yaml:"wildcards"`                                     // List of wildcards to search projects from
+		// Token to use to authenticate against the API
+		Token string `yaml:"token"`
+
+		// The URL of the GitLab server/api health endpoint (default to /users/sign_in which is publicly available on gitlab.com)
+		HealthURL string `yaml:"health_url"`
+
+		// Whether to validate the service is reachable calling HealthURL
+		DisableHealthCheck bool `yaml:"disable_health_check"`
+
+		// Whether to skip TLS validation when querying HealthURL
+		DisableTLSVerify bool `yaml:"disable_tls_verify"`
+	} `yaml:"gitlab"`
+
+	// Maximum amount of requests per seconds to make against the GitLab API (default: 10)
+	MaximumGitLabAPIRequestsPerSecond int `yaml:"maximum_gitlab_api_requests_per_second"`
+
+	// Interval in seconds at which to poll projects from wildcards
+	ProjectsPollingIntervalSeconds int `yaml:"projects_polling_interval_seconds"`
+
+	// Interval in seconds to fetch refs from projects
+	RefsPollingIntervalSeconds int `yaml:"refs_polling_interval_seconds"`
+
+	// Interval in seconds to get new pipelines from refs (exponentially backing of to maximum value)
+	PipelinesPollingIntervalSeconds int `yaml:"pipelines_polling_interval_seconds"`
+
+	// Whether to attempt retrieving refs from pipelines when the exporter starts
+	OnInitFetchRefsFromPipelines bool `yaml:"on_init_fetch_refs_from_pipelines"`
+
+	// Maximum number of pipelines to analyze per project to search for refs on init (default: 100)
+	OnInitFetchRefsFromPipelinesDepthLimit int `yaml:"on_init_fetch_refs_from_pipelines_depth_limit"`
+
+	// Sets the parallelism for polling projects from the API
+	MaximumProjectsPollingWorkers int `yaml:"maximum_projects_poller_workers"`
+
+	// Disable OpenMetrics content encoding in prometheus HTTP handler (default: false)
+	DisableOpenmetricsEncoding bool `yaml:"disable_openmetrics_encoding"`
+
+	// Default parameters which can be overridden at either the Project or Wildcard level
+	Defaults Parameters `yaml:"defaults"`
+
+	// List of projects to poll
+	Projects []Project `yaml:"projects"`
+
+	// List of wildcards to search projects from
+	Wildcards []Wildcard `yaml:"wildcards"`
+}
+
+// Parameters for the fetching configuration of Projects and Wildcards
+type Parameters struct {
+	// Whether to attempt to retrieve job metrics from polled pipelines
+	FetchPipelineJobMetricsValue *bool `yaml:"fetch_pipeline_job_metrics"`
+
+	// Whether to report all pipeline / job statuses, or only report the one from the last job.
+	OutputSparseStatusMetricsValue *bool `yaml:"output_sparse_status_metrics"`
+
+	// Whether to attempt to retrieve variables included in the pipeline execution
+	FetchPipelineVariablesValue *bool `yaml:"fetch_pipeline_variables"`
+
+	// Regular expression to filter pipeline variables values to fetch (defaults to '.*')
+	PipelineVariablesRegexpValue *string `yaml:"pipeline_variables_regexp"`
+
+	// Regular expression to filter project refs to fetch (defaults to '.*')
+	RefsRegexpValue *string `yaml:"refs_regexp"`
 }
 
 // Project holds information about a GitLab project
 type Project struct {
-	Name                      string `yaml:"name"`
-	RefsRegexp                string `yaml:"refs_regexp"`
-	FetchPipelineJobMetrics   *bool  `yaml:"fetch_pipeline_job_metrics,omitempty"`
-	OutputSparseStatusMetrics *bool  `yaml:"output_sparse_status_metrics,omitempty"`
+	Parameters `yaml:",inline"`
+
+	Name string `yaml:"name"`
 }
 
 // Wildcard is a specific handler to dynamically search projects
 type Wildcard struct {
+	Parameters `yaml:",inline"`
+
 	Search string `yaml:"search"`
 	Owner  struct {
 		Name             string `yaml:"name"`
 		Kind             string `yaml:"kind"`
 		IncludeSubgroups bool   `yaml:"include_subgroups"`
-	}
-	Archived                  bool   `yaml:"archived"`
-	FetchPipelineJobMetrics   *bool  `yaml:"fetch_pipeline_job_metrics,omitempty"`
-	OutputSparseStatusMetrics *bool  `yaml:"output_sparse_status_metrics,omitempty"`
-	RefsRegexp                string `yaml:"refs_regexp"`
+	} `yaml:"owner"`
+	Archived bool `yaml:"archived"`
 }
 
 // Default values
@@ -66,30 +108,81 @@ const (
 	defaultProjectsPollingIntervalSeconds         = 1800
 	defaultRefsPollingIntervalSeconds             = 300
 
+	defaultFetchPipelineJobMetrics   = false
+	defaultOutputSparseStatusMetrics = false
+	defaultFetchPipelineVariables    = false
+	defaultRefsRegexp                = `^master$`
+	defaultPipelineVariablesRegexp   = `.*`
+
 	errNoProjectsOrWildcardConfigured = "you need to configure at least one project/wildcard to poll, none given"
 	errConfigFileNotFound             = "couldn't open config file : %v"
-
-	variablesCatchallRegex = "\\.*"
 )
 
 var cfg = &Config{}
 
-// ShouldFetchPipelineJobMetrics returns true if pipeline job statistics should be fetched
-func (p *Project) ShouldFetchPipelineJobMetrics(cfg *Config) bool {
-	if p.FetchPipelineJobMetrics == nil {
-		// Default to global config value
-		return cfg.FetchPipelineJobMetrics
+// FetchPipelineJobMetrics ...
+func (p *Project) FetchPipelineJobMetrics(cfg *Config) bool {
+	if p.FetchPipelineJobMetricsValue != nil {
+		return *p.FetchPipelineJobMetricsValue
 	}
-	return *p.FetchPipelineJobMetrics
+
+	if cfg.Defaults.FetchPipelineJobMetricsValue != nil {
+		return *cfg.Defaults.FetchPipelineJobMetricsValue
+	}
+
+	return defaultFetchPipelineJobMetrics
 }
 
-// ShouldOutputSparseStatusMetrics returns true if sparse status metrics should be exported
-func (p *Project) ShouldOutputSparseStatusMetrics(cfg *Config) bool {
-	if p.OutputSparseStatusMetrics == nil {
-		// Default to global config value
-		return cfg.OutputSparseStatusMetrics
+// OutputSparseStatusMetrics ...
+func (p *Project) OutputSparseStatusMetrics(cfg *Config) bool {
+	if p.OutputSparseStatusMetricsValue != nil {
+		return *p.OutputSparseStatusMetricsValue
 	}
-	return *p.OutputSparseStatusMetrics
+
+	if cfg.Defaults.OutputSparseStatusMetricsValue != nil {
+		return *cfg.Defaults.OutputSparseStatusMetricsValue
+	}
+
+	return defaultOutputSparseStatusMetrics
+}
+
+// FetchPipelineVariables ...
+func (p *Project) FetchPipelineVariables(cfg *Config) bool {
+	if p.FetchPipelineVariablesValue != nil {
+		return *p.FetchPipelineVariablesValue
+	}
+
+	if cfg.Defaults.FetchPipelineVariablesValue != nil {
+		return *cfg.Defaults.FetchPipelineVariablesValue
+	}
+
+	return defaultFetchPipelineVariables
+}
+
+// PipelineVariablesRegexp ...
+func (p *Project) PipelineVariablesRegexp(cfg *Config) string {
+	if p.PipelineVariablesRegexpValue != nil {
+		return *p.PipelineVariablesRegexpValue
+	}
+
+	if cfg.Defaults.PipelineVariablesRegexpValue != nil {
+		return *cfg.Defaults.PipelineVariablesRegexpValue
+	}
+
+	return defaultPipelineVariablesRegexp
+}
+
+// RefsRegexp ...
+func (p *Project) RefsRegexp(cfg *Config) string {
+	if p.RefsRegexpValue != nil {
+		return *p.RefsRegexpValue
+	}
+
+	if cfg.Defaults.RefsRegexpValue != nil {
+		return *cfg.Defaults.RefsRegexpValue
+	}
+
+	return defaultRefsRegexp
 }
 
 // Parse loads a yaml file into a Config structure
@@ -141,18 +234,5 @@ func (cfg *Config) Parse(path string) error {
 		cfg.MaximumProjectsPollingWorkers = runtime.GOMAXPROCS(0)
 	}
 
-	if cfg.PipelineVariablesFilterRegexp == "" {
-		cfg.PipelineVariablesFilterRegexp = variablesCatchallRegex
-	}
-
 	return nil
-}
-
-// MergeWithContext is used to override values defined in the config by ones
-// provided at runtime
-func (cfg *Config) MergeWithContext(ctx *cli.Context) {
-	token := ctx.GlobalString("gitlab-token")
-	if len(token) != 0 {
-		cfg.Gitlab.Token = token
-	}
 }